moves packages dataset and resources to dimcat.data
