+++ src/dimcat/data/resources/results.py
