+++ src/dimcat/steps/analyzers/counters.py
@@ -1,7 +1,7 @@
 import logging
 
-from dimcat.resources.base import DimcatResource, SomeDataframe, SomeSeries
-from dimcat.resources.features import Feature
+from dimcat.data.resources.base import DimcatResource, SomeDataframe, SomeSeries
+from dimcat.data.resources.features import Feature
 from dimcat.steps.analyzers.base import Analyzer
 
 logger = logging.getLogger(__name__)
