+++ src/dimcat/steps/groupers/base.py
@@ -2,10 +2,10 @@ import logging
 from typing import Dict, List, Sequence
 
 import pandas as pd
-from dimcat.dataset.processed import GroupedDataset
-from dimcat.resources.base import DimcatIndex, IndexField
-from dimcat.resources.features import Feature
-from dimcat.resources.utils import make_index_from_grouping_dict
+from dimcat.data.dataset.processed import GroupedDataset
+from dimcat.data.resources.base import DimcatIndex, IndexField
+from dimcat.data.resources.features import Feature
+from dimcat.data.resources.utils import make_index_from_grouping_dict
 from dimcat.steps import PipelineStep
 from dimcat.utils import check_name
 from marshmallow import fields
