+++ tests/test_grouper.py
@@ -1,7 +1,7 @@
 from random import sample
 
 import pytest
-from dimcat.dataset.processed import GroupedDataset
+from dimcat.data.dataset.processed import GroupedDataset
 from dimcat.steps.groupers.base import CustomPieceGrouper
